+++ database/migrations/2016_12_22_102218_create_tags_table.php
@@ -21,8 +21,8 @@ class CreateTagsTable extends Migration
             $table->string('slug');
             $table->{$this->jsonable()}('name');
             $table->{$this->jsonable()}('description')->nullable();
-            $table->integer('order')->default(0);
-            $table->string('type')->nullable();
+            $table->mediumInteger('sort_order')->unsigned()->default(0);
+            $table->string('group')->nullable();
             $table->timestamps();
             $table->softDeletes();