+++ database/migrations/2016_12_22_105224_create_taggables_table.php
@@ -17,14 +17,14 @@ class CreateTaggablesTable extends Migration
     {
         Schema::create(config('rinvex.taggable.tables.taggables'), function (Blueprint $table) {
             // Columns
-            $table->unsignedInteger('tag_id');
-            $table->unsignedInteger('taggable_id');
+            $table->integer('tag_id')->unsigned();
+            $table->integer('taggable_id')->unsigned();
             $table->string('taggable_type');
             $table->timestamps();
 
             // Indexes
             $table->unique(['tag_id', 'taggable_id', 'taggable_type'], 'taggables_ids_type_unique');
-            $table->foreign('tag_id')->references('id')->on('tags')
+            $table->foreign('tag_id')->references('id')->on(config('rinvex.taggable.tables.tags'))
                   ->onDelete('cascade')->onUpdate('cascade');
         });
     }