+++ CHANGELOG.md
@@ -5,5 +5,18 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](CONTRIBUTING.md).
 
 
+## [v0.0.2] - 2017-06-29
+- Enforce consistency
+- Update validation rules
+- Add Laravel 5.5 support
+- Fix detaching tags method
+- Fix multiple bugs & issues
+- Change integer column length
+- Tweak model event registration
+- Rename tags type to tags group
+- Enforce more secure approach using model fillable instead of guarded
+
 ## v0.0.1 - 2017-04-08
 - Rename package to "rinvex/taggable" from "rinvex/tag" based on 3b6a727
+
+[v0.0.5]: https://github.com/rinvex/taggable/compare/v0.0.1...v0.0.2